enhance go fileperm rule
Relaxes the minimum permission mask by allowing group read which is a
common setting. It will still flag on any permissions for Other/world as
giving read permissions for any user on the system on a file could
definitely be a security issue in some circumstances.

Also adds `os.WriteFile` to the list of patterns as this function also
allows to set file permissions when creating a file.